+++ packages/mds-jurisdiction/handlers/create-jurisdiction.ts
@@ -20,7 +20,7 @@ import {
   JurisdictionDomainModel
 } from '@mds-core/mds-jurisdiction-service'
 import { HandleServiceResponse } from '@mds-core/mds-service-helpers'
-import { JurisdictionApiRequest, JurisdictionApiResponse } from '../types'
+import { JurisdictionApiRequest, JurisdictionApiResponse } from '../@types'
 
 interface CreateJurisdictionRequest extends JurisdictionApiRequest {
   body: CreateJurisdictionDomainModel | CreateJurisdictionDomainModel[]
