+++ packages/mds-jurisdiction/handlers/get-jurisdictions.ts
@@ -19,7 +19,7 @@ import { HandleServiceResponse } from '@mds-core/mds-service-helpers'
 import { parseRequest } from '@mds-core/mds-api-helpers'
 import { ApiQuery } from '@mds-core/mds-api-server'
 import { HasJurisdictionClaim } from './utils'
-import { JurisdictionApiRequest, JurisdictionApiResponse } from '../types'
+import { JurisdictionApiRequest, JurisdictionApiResponse } from '../@types'
 
 type GetJurisdictionsRequest = JurisdictionApiRequest & ApiQuery<'effective'>
