+++ packages/mds-jurisdiction/handlers/update-jurisdiction.ts
@@ -21,7 +21,7 @@ import {
   JurisdictionIdType
 } from '@mds-core/mds-jurisdiction-service'
 import { HandleServiceResponse } from '@mds-core/mds-service-helpers'
-import { JurisdictionApiRequest, JurisdictionApiResponse } from '../types'
+import { JurisdictionApiRequest, JurisdictionApiResponse } from '../@types'
 
 interface UpdateJurisdictionRequest extends JurisdictionApiRequest<{ jurisdiction_id: JurisdictionIdType }> {
   body: UpdateJurisdictionDomainModel
