+++ packages/mds-jurisdiction/handlers/utils.ts
@@ -15,7 +15,7 @@
  */
 
 import { JurisdictionDomainModel } from '@mds-core/mds-jurisdiction-service'
-import { JurisdictionApiResponse } from '../types'
+import { JurisdictionApiResponse } from '../@types'
 
 export const HasJurisdictionClaim = <TBody extends {}>(res: JurisdictionApiResponse<TBody>) => (
   jurisdiction: JurisdictionDomainModel
