+++ packages/mds-jurisdiction/middleware/jurisdiction-api-version.ts
@@ -15,7 +15,7 @@
  */
 
 import { ApiVersionMiddleware } from '@mds-core/mds-api-server'
-import { JURISDICTION_API_SUPPORTED_VERSIONS, JURISDICTION_API_DEFAULT_VERSION } from '../types'
+import { JURISDICTION_API_SUPPORTED_VERSIONS, JURISDICTION_API_DEFAULT_VERSION } from '../@types'
 
 export const JurisdictionApiVersionMiddleware = ApiVersionMiddleware(
   'application/vnd.mds.jurisdiction+json',
